requirements.txt
aiohttp
earthengine_api
folium
geopy
langchain
langchain_google_genai
numpy
pandas
Pillow
plotly
pvlib
pydeck
pyproj
pytz
Requests
Shapely
streamlit==1.37.1
streamlit_browser_session_storage
streamlit_drawable_canvas
streamlit_extras
streamlit_folium
